Add files via upload
GitHub Repository Name: "GuessTheNumber-CSharp"

Description:
The GitHub repository "GuessTheNumber-CSharp" contains a C# project that allows playing a number guessing game. The program randomly generates a number between 0 and 1000 and prompts the player to guess the number. The player can enter numbers, and the program provides hints to guide the player towards the correct answer. Once the player finds the correct number, the program displays a congratulatory message. This project serves as a simple demonstration of using loops, conditions, and input/output in C#.